+++ tests/dummy/app/router.js
@@ -7,7 +7,9 @@ const DummyRouter = Router.extend({
 });
 
 DummyRouter.map(function() {
-
+  this.route('simple');
+  this.route('reference');
+  this.route('position');
 });
 
 export default DummyRouter;