+++ tests/helpers/utils.js
@@ -1,5 +1,8 @@
 export function positionClasses(tooltip) {
-  return tooltip.getAttribute('class').split(/\s+/).filter(name => {
-    return /is-(north|south|east|west)/.test(name);
-  });
+  return tooltip
+    .getAttribute('class')
+    .split(/\s+/)
+    .filter(name => {
+      return /is-(north|south|east|west)/.test(name);
+    });
 }