+++ tests/integration/components/render-active-tooltips-test.js
@@ -1,6 +1,6 @@
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
-import { render, settled, findAll, triggerEvent, click } from '@ember/test-helpers';
+import { render, settled, triggerEvent, click } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 import TooltipComponent from '@zestia/ember-async-tooltips/components/tool-tip';
 import TooltipperComponent from '@zestia/ember-async-tooltips/components/tool-tipper';
@@ -11,10 +11,13 @@ module('render-active-tooltips', function(hooks) {
   test('it renders', async function(assert) {
     assert.expect(1);
 
-    await render(hbs`{{render-active-tooltips}}`);
+    await render(hbs`<RenderActiveTooltips />`);
 
-    assert.equal(this.get('element').innerHTML, '<!---->',
-      'does not blow up if no tooltips are active');
+    assert.equal(
+      this.get('element').innerHTML,
+      '<!---->',
+      'does not blow up if no tooltips are active'
+    );
   });
 
   test('it renders tooltip components', async function(assert) {
@@ -23,7 +26,7 @@ module('render-active-tooltips', function(hooks) {
     const FooTooltipComponent = TooltipComponent.extend({
       classNames: ['foo-tooltip'],
       layout: hbs`
-        Hello World ({{@myAttr}})<br>
+        Hello World ({{@myArg}})<br>
         <button class="hide-from-tooltip" onclick={{action "hide"}}></button>
       `
     });
@@ -32,64 +35,60 @@ module('render-active-tooltips', function(hooks) {
       classNames: ['foo-tooltipper']
     });
 
-    this.owner.register('component:foo-tooltip',    FooTooltipComponent);
+    this.owner.register('component:foo-tooltip', FooTooltipComponent);
     this.owner.register('component:foo-tooltipper', FooTooltipperComponent);
 
     await render(hbs`
       <div class="in">
-        {{#foo-tooltipper tooltip=(component "foo-tooltip" myMattr="foo") as |tt|}}
+        <FooTooltipper @tooltip={{component "foo-tooltip" myArg="foo"}} as |tt|>
           <button class="show-from-tooltipper" onclick={{action tt.showTooltip}}></button>
           <button class="hide-from-tooltipper" onclick={{action tt.hideTooltip}}></button>
-        {{/foo-tooltipper}}
+        </FooTooltipper>
       </div>
 
       <div class="out">
-        {{render-active-tooltips}}
+        <RenderActiveTooltips />
       </div>
     `);
 
-    triggerEvent('.in .foo-tooltipper', 'mouseover');
+    triggerEvent('.in .foo-tooltipper', 'mouseenter');
 
-    assert.equal(findAll('.foo-tooltip').length, 0,
-      'tooltip is not rendered yet (still hovering over it)');
+    assert.dom('.foo-tooltip').doesNotExist('tooltip is not rendered yet (still hovering over it)');
 
     await settled();
 
-    assert.equal(findAll('.out .foo-tooltip').length, 1,
-      'the tooltip is rendered elsewhere');
+    assert.dom('.out .foo-tooltip').exists({ count: 1 }, 'the tooltip is rendered elsewhere');
 
-    await triggerEvent('.foo-tooltipper', 'mouseout');
+    await triggerEvent('.foo-tooltipper', 'mouseleave');
 
-    assert.equal(findAll('.foo-tooltip').length, 1,
-      'tooltip is not destroyed yet (due to hover delay)');
+    assert
+      .dom('.foo-tooltip')
+      .exists({ count: 1 }, 'tooltip is not destroyed yet (due to hover delay)');
 
     await triggerEvent('.foo-tooltip', 'animationEnd');
 
-    assert.equal(findAll('.out .foo-tooltip').length, 0,
-      'the tooltip is destroyed after its hide animation');
+    assert
+      .dom('.out .foo-tooltip')
+      .doesNotExist('the tooltip is destroyed after its hide animation');
 
     await click('.show-from-tooltipper');
 
-    assert.equal(findAll('.foo-tooltip').length, 1,
-      'tooltip can be manually shown by tooltipper');
+    assert.dom('.foo-tooltip').exists({ count: 1 }, 'tooltip can be manually shown by tooltipper');
 
     await click('.hide-from-tooltipper');
 
     await triggerEvent('.foo-tooltip', 'animationEnd');
 
-    assert.equal(findAll('.foo-tooltip').length, 0,
-      'tooltip can be manually hidden by tooltipper');
+    assert.dom('.foo-tooltip').doesNotExist('tooltip can be manually hidden by tooltipper');
 
-    await triggerEvent('.foo-tooltipper', 'mouseover');
+    await triggerEvent('.foo-tooltipper', 'mouseenter');
 
-    assert.equal(findAll('.foo-tooltip').length, 1,
-      'precondition: tooltip shown');
+    assert.dom('.foo-tooltip').exists({ count: 1 }, 'precondition: tooltip shown');
 
     await click('.hide-from-tooltip');
 
     await triggerEvent('.foo-tooltip', 'animationEnd');
 
-    assert.equal(findAll('.foo-tooltip').length, 0,
-      'tooltip can be hidden by itself');
+    assert.dom('.foo-tooltip').doesNotExist('tooltip can be hidden by itself');
   });
 });