+++ tests/integration/components/tool-tipper-test.js
@@ -1,6 +1,6 @@
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
-import { settled, render, triggerEvent, find, findAll } from '@ember/test-helpers';
+import { settled, render, triggerEvent, find } from '@ember/test-helpers';
 import { defer } from 'rsvp';
 import hbs from 'htmlbars-inline-precompile';
 
@@ -10,16 +10,15 @@ module('tool-tipper', function(hooks) {
   test('it renders', async function(assert) {
     assert.expect(3);
 
-    await render(hbs`{{tool-tipper}}`);
+    await render(hbs`<ToolTipper />`);
 
-    assert.equal(findAll('.tooltipper').length, 1,
-      'tool-tipper components have an appropriate class name');
+    assert
+      .dom('.tooltipper')
+      .exists({ count: 1 }, 'tool-tipper components have an appropriate class name');
 
-    assert.equal(find('.tooltipper').tagName, 'SPAN',
-      'renders as an inline element by default');
+    assert.equal(find('.tooltipper').tagName, 'SPAN', 'renders as an inline element by default');
 
-    assert.strictEqual(find('.tooltipper').getAttribute('type'), null,
-      'is not a button');
+    assert.strictEqual(find('.tooltipper').getAttribute('type'), null, 'is not a button');
   });
 
   test('as a hyperlink', async function(assert) {
@@ -27,14 +26,11 @@ module('tool-tipper', function(hooks) {
 
     await render(hbs`{{tool-tipper tagName="a" href="foo" rel="bar" target="baz"}}`);
 
-    assert.equal(find('.tooltipper').getAttribute('href'), 'foo',
-      'can set href attribute');
+    assert.dom('.tooltipper').hasAttribute('href', 'foo', 'can set href attribute');
 
-    assert.equal(find('.tooltipper').getAttribute('rel'), 'bar',
-      'can set rel attribute');
+    assert.dom('.tooltipper').hasAttribute('rel', 'bar', 'can set rel attribute');
 
-    assert.equal(find('.tooltipper').getAttribute('target'), 'baz',
-      'can set target attribute');
+    assert.dom('.tooltipper').hasAttribute('target', 'baz', 'can set target attribute');
   });
 
   test('as a button', async function(assert) {
@@ -42,27 +38,29 @@ module('tool-tipper', function(hooks) {
 
     await render(hbs`{{tool-tipper tagName="button"}}`);
 
-    assert.equal(find('.tooltipper').getAttribute('type'), 'button',
-      'renders as a non-submittion button by default');
+    assert
+      .dom('.tooltipper')
+      .hasAttribute('type', 'button', 'renders as a non-submittion button by default');
 
     await render(hbs`{{tool-tipper tagName="button" type="foo"}}`);
 
-    assert.equal(find('.tooltipper').getAttribute('type'), 'foo',
-      'can specify type of button');
+    assert.dom('.tooltipper').hasAttribute('type', 'foo', 'can specify type of button');
   });
 
   test('tabindex', async function(assert) {
     assert.expect(2);
 
-    await render(hbs`{{tool-tipper}}`);
+    await render(hbs`<ToolTipper />`);
 
-    assert.strictEqual(find('.tooltipper').getAttribute('tabindex'), null,
-      'no default tabindex');
+    assert.strictEqual(find('.tooltipper').getAttribute('tabindex'), null, 'no default tabindex');
 
-    await render(hbs`{{tool-tipper tabindex=-1}}`);
+    await render(hbs`{{tool-tipper tabindex="-1"}}`);
 
-    assert.strictEqual(find('.tooltipper').getAttribute('tabindex'), '-1',
-      'can set tabindex attribute');
+    assert.strictEqual(
+      find('.tooltipper').getAttribute('tabindex'),
+      '-1',
+      'can set tabindex attribute'
+    );
   });
 
   test('onLoad action', async function(assert) {
@@ -77,23 +75,23 @@ module('tool-tipper', function(hooks) {
       return deferred.promise;
     });
 
-    await render(hbs`{{tool-tipper onLoad=(action load)}}`);
+    await render(hbs`<ToolTipper @onLoad={{action this.load}} />`);
 
-    await triggerEvent('.tooltipper', 'mouseover');
+    await triggerEvent('.tooltipper', 'mouseenter');
 
-    assert.ok(find('.tooltipper').classList.contains('is-loading'),
-      'has a loading class whilst loading the data');
+    assert.dom('.tooltipper').hasClass('is-loading', 'has a loading class whilst loading the data');
 
     deferred.resolve();
 
     await settled();
 
-    assert.ok(!find('.tooltipper').classList.contains('is-loading'),
-      'loading class is removed when loading is complete');
+    assert.ok(
+      !find('.tooltipper').classList.contains('is-loading'),
+      'loading class is removed when loading is complete'
+    );
 
-    await triggerEvent('.tooltipper', 'mouseover');
+    await triggerEvent('.tooltipper', 'mouseenter');
 
-    assert.equal(count, 1,
-      'only loads once');
+    assert.equal(count, 1, 'only loads once');
   });
 });